// src/types/userAndSkills.ts
export interface UserAndSkills {
    id: string;
    email: string;
    firstName: string;
    lastName: string;
    avatar: string;
    courseCompletion: number;
    courseEngagement: number;
    projectDegree: number;
    teamProjectDegree: number;
    expectedTypeWork: string;
    targetWorkCity: string;
    expectedContractType: string;
    expectedSalary: string;
    canTakeApprenticeship: string;
    monthsOfCommercialExp: number;
    reservationExpiryDate: string;
}